Unify prop utils
