+++ src/components/avatar/avatar.js
@@ -1,8 +1,8 @@
 import Vue from '../../utils/vue'
-import pluckProps from '../../utils/pluck-props'
 import { getComponentConfig } from '../../utils/config'
 import { isNumber, isString, isUndefinedOrNull } from '../../utils/inspect'
 import { toFloat } from '../../utils/number'
+import { pluckProps } from '../../utils/props'
 import { BButton } from '../button/button'
 import { BLink, props as BLinkProps } from '../link/link'
 import { BIcon } from '../../icons/icon'
