+++ src/components/badge/badge.js
@@ -1,8 +1,8 @@
-import Vue from '../../utils/vue'
-import pluckProps from '../../utils/pluck-props'
 import { mergeData } from 'vue-functional-data-merge'
+import Vue from '../../utils/vue'
 import { getComponentConfig } from '../../utils/config'
 import { clone } from '../../utils/object'
+import { pluckProps } from '../../utils/props'
 import { BLink, props as BLinkProps } from '../link/link'
 
 const NAME = 'BBadge'
