+++ src/components/button/button.js
@@ -1,12 +1,12 @@
-import Vue from '../../utils/vue'
 import { mergeData } from 'vue-functional-data-merge'
+import Vue from '../../utils/vue'
 import KeyCodes from '../../utils/key-codes'
-import pluckProps from '../../utils/pluck-props'
 import { concat } from '../../utils/array'
 import { getComponentConfig } from '../../utils/config'
 import { addClass, removeClass } from '../../utils/dom'
 import { isBoolean, isEvent, isFunction } from '../../utils/inspect'
 import { clone } from '../../utils/object'
+import { pluckProps } from '../../utils/props'
 import { toString } from '../../utils/string'
 import { BLink, props as BLinkProps } from '../link/link'
