+++ src/components/card/card-footer.js
@@ -1,8 +1,7 @@
 import { mergeData } from 'vue-functional-data-merge'
 import Vue from '../../utils/vue'
-import copyProps from '../../utils/copy-props'
-import prefixPropName from '../../utils/prefix-prop-name'
 import { htmlOrText } from '../../utils/html'
+import { copyProps, prefixPropName } from '../../utils/props'
 import cardMixin from '../../mixins/card'
 
 // --- Props ---
