+++ src/components/layout/col.js
@@ -1,11 +1,11 @@
 import { mergeData } from 'vue-functional-data-merge'
 import identity from '../../utils/identity'
 import memoize from '../../utils/memoize'
-import suffixPropName from '../../utils/suffix-prop-name'
 import { arrayIncludes } from '../../utils/array'
 import { getBreakpointsUpCached } from '../../utils/config'
 import { isUndefinedOrNull } from '../../utils/inspect'
 import { assign, create, keys } from '../../utils/object'
+import { suffixPropName } from '../../utils/props'
 import { lowerCase } from '../../utils/string'
 
 const RX_COL_CLASS = /^col-/
