+++ src/components/layout/row.js
@@ -1,10 +1,10 @@
 import { mergeData } from 'vue-functional-data-merge'
 import identity from '../../utils/identity'
 import memoize from '../../utils/memoize'
-import suffixPropName from '../../utils/suffix-prop-name'
 import { arrayIncludes, concat } from '../../utils/array'
 import { getBreakpointsUpCached } from '../../utils/config'
 import { create, keys } from '../../utils/object'
+import { suffixPropName } from '../../utils/props'
 import { lowerCase, toString, trim } from '../../utils/string'
 
 const COMMON_ALIGNMENT = ['start', 'end', 'center']
