+++ src/components/link/link.js
@@ -1,8 +1,8 @@
 import Vue from '../../utils/vue'
-import pluckProps from '../../utils/pluck-props'
 import { concat } from '../../utils/array'
 import { attemptBlur, attemptFocus } from '../../utils/dom'
 import { isBoolean, isEvent, isFunction, isUndefined } from '../../utils/inspect'
+import { pluckProps } from '../../utils/props'
 import { computeHref, computeRel, computeTag, isRouterLink } from '../../utils/router'
 import attrsMixin from '../../mixins/attrs'
 import listenersMixin from '../../mixins/listeners'
