+++ src/components/list-group/list-group-item.js
@@ -1,9 +1,9 @@
 import { mergeData } from 'vue-functional-data-merge'
 import Vue from '../../utils/vue'
-import pluckProps from '../../utils/pluck-props'
 import { arrayIncludes } from '../../utils/array'
 import { getComponentConfig } from '../../utils/config'
 import { clone } from '../../utils/object'
+import { pluckProps } from '../../utils/props'
 import { BLink, props as BLinkProps } from '../link/link'
 
 // --- Constants ---
