+++ src/components/navbar/navbar-nav.js
@@ -1,6 +1,6 @@
-import Vue from '../../utils/vue'
 import { mergeData } from 'vue-functional-data-merge'
-import pluckProps from '../../utils/pluck-props'
+import Vue from '../../utils/vue'
+import { pluckProps } from '../../utils/props'
 import { props as BNavProps } from '../nav/nav'
 
 // -- Constants --
