+++ src/components/pagination-nav/pagination-nav.js
@@ -1,12 +1,12 @@
 import Vue from '../../utils/vue'
 import looseEqual from '../../utils/loose-equal'
-import pluckProps from '../../utils/pluck-props'
 import { getComponentConfig } from '../../utils/config'
 import { attemptBlur, requestAF } from '../../utils/dom'
 import { isBrowser } from '../../utils/env'
 import { isArray, isUndefined, isFunction, isObject } from '../../utils/inspect'
 import { mathMax } from '../../utils/math'
 import { toInteger } from '../../utils/number'
+import { pluckProps } from '../../utils/props'
 import { computeHref, parseQuery } from '../../utils/router'
 import { toString } from '../../utils/string'
 import { warn } from '../../utils/warn'
