--- src/utils/prefix-prop-name.js
@@ -1,9 +1,0 @@
-import { upperFirst } from './string'
-
-/**
- * @param {string} prefix
- * @param {string} value
- */
-const prefixPropName = (prefix, value) => prefix + upperFirst(value)
-
-export default prefixPropName
